Migrate app.php to IBootstrap
